Remove usage of PoiFragment and pass everything into  Poi Activity
