+++ app/src/main/java/com/github/epfl/meili/messages/ChatLogActivity.kt
@@ -14,7 +14,6 @@ import com.github.epfl.meili.map.MapActivity
 import com.github.epfl.meili.models.ChatMessage
 import com.github.epfl.meili.models.PointOfInterest
 import com.github.epfl.meili.models.User
-import com.github.epfl.meili.poi.PoiActivity
 import com.github.epfl.meili.profile.friends.FriendsListActivity.Companion.FRIEND_KEY
 import com.github.epfl.meili.util.DateAuxiliary
 import com.github.epfl.meili.util.MenuActivity
@@ -132,8 +131,8 @@ class ChatLogActivity : MenuActivity(R.menu.nav_chat_menu) {
         val groupMessageObserver = Observer<List<ChatMessage>?> { list ->
             val newMessages = list.minus(messageSet)
             newMessages.filter { message -> message.toId == chatID }.forEach { message ->
-                    Log.d(TAG, "loading message: ${message.text}")
-                    adapter.add(ChatItem(message, message.fromId == currentUser!!.uid, isGroupChat))
+                Log.d(TAG, "loading message: ${message.text}")
+                adapter.add(ChatItem(message, message.fromId == currentUser!!.uid, isGroupChat))
             }
 
             messageSet.addAll(newMessages)
